+++ common/src/main/java/com/iamkaf/bonded/registry/TierMap.java
@@ -227,6 +227,9 @@ public class TierMap {
         addExperienceCap(Items.BOW, 150);
         addRepairMaterial(Items.BOW, Items.STRING);
 
+        addExperienceCap(Items.CROSSBOW, 150);
+        addRepairMaterial(Items.CROSSBOW, Items.STRING);
+
         addExperienceCap(Items.BRUSH, 150);
         addRepairMaterial(Items.BRUSH, Items.FEATHER);